+++ addon/routes/edit-form-new.js
@@ -63,14 +63,16 @@ export default EditFormRoute.extend({
       flexberryDetailInteractionService.set('modelSelectedDetail', undefined);
 
       if (modelCurrentNotSaved) {
-        return modelCurrentNotSaved;
+        return this.returnNewModel(modelCurrentNotSaved);
       }
 
       if (modelSelectedDetail) {
-        return modelSelectedDetail;
+        return this.returnNewModel(modelSelectedDetail);
       }
 
-      return store.createRecord(modelName, { id: generateUniqueId() });
+      let model = store.createRecord(modelName, { id: generateUniqueId() });
+
+      return this.returnNewModel(model);
     }
 
     // Get the copyable instance.
@@ -85,6 +87,16 @@ export default EditFormRoute.extend({
 
       return record;
     });
+  },
+
+  /**
+    Return model as Primese.
+
+    @method returnNewModel
+    @param {Object} model
+   */
+  returnNewModel(value) {
+    return new Ember.RSVP.resolve(value);
   },
 
   /**