+++ blueprints/flexberry-common/CommonUtils.js
@@ -36,7 +36,10 @@ var CommonUtils = (function () {
     };
     CommonUtils.getFilesForGeneration = function (blueprint, exclusiveFunction) {
         if (exclusiveFunction === void 0) { exclusiveFunction = null; }
-        var files = Blueprint.prototype.files.call(blueprint);
+        blueprint._super._files = null;
+        blueprint._super.path = blueprint.path;
+        var files = blueprint._super.files.apply(blueprint);
+        blueprint._super._files = null;
         if (exclusiveFunction != null) {
             lodash.remove(files, exclusiveFunction);
         }