+++ tests/acceptance/edit-form-validation-test/validation-test.js
@@ -14,22 +14,7 @@ executeTest('check complete all tests', (store, assert, app) => {
   andThen(() => {
     assert.equal(currentPath(), path);
 
-    let $validationDataField = $('.flexberry-simpledatetime');
-    let $validationDataDiv = $validationDataField.children('.input');
-    let $validationDataInput = $validationDataDiv.children('.flatpickr-input');
-
-    run(() => {
-      // Open datepicker calendar.
-      $($validationDataInput[0]).click();
-      let $validationDateButton = $('.flatpickr-day')[16];
-
-      // Select date.
-      click($validationDateButton);
-    });
-
-    let $validationFlexberryLookup = $('.flexberry-lookup');
-    let $validationFlexberryLookupInput = $validationFlexberryLookup.children('.input');
-    let $validationFlexberryLookupButton = $validationFlexberryLookupInput.children('.ui-change.button');
+    let $validationFlexberryLookupButton = Ember.$('.ui.button.ui-change')[0];
 
     // Click lookup button.
     run(() => {