+++ tests/dummy/app/app.js
@@ -2,6 +2,8 @@ import Application from '@ember/application';
 import Resolver from './resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from './config/environment';
+import './default-user-settings/index';
+import './models/custom-inflector-rules';
 
 const App = Application.extend({
   modulePrefix: config.modulePrefix,