+++ CHANGELOG.md
@@ -4,32 +4,51 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-## [3.6.0-beta.1] - 2021-04-29
-### Changed
-- Add some improvements for UI of generated application.
-
-## [3.6.0-beta.0] - 2020-10-07
-### Added
-- Changes from [`2.6.0-beta.0`](#260-beta0---2020-10-06) version.
+## [2.6.1] - 2021-07-02
+### Fixed
+* The `flexberry-model` blueprint generated wrong model mixin when project contains `ember-cp-validations` dependency.
 
 ### Changed
-* Update dependency on `ember-flexberry-data` to version `3.6.0-beta.0`.
+* Update dependency on `ember-flexberry-data` to version `2.6.0`.
 
-## [2.6.0-beta.0] - 2020-10-06
+## [2.6.0] - 2021-06-16
 ### Added
-* The `lookupOnChange` event in the `LookupEvents` service, called when the value in the `flexberry-lookup` component changes.
-* Support for the `flexberry-lookup` component block form, which was previously only available in the mobile version.
+* The `flexberry-file` component:
+    * Added upload/download file options for mobile component's menu.
+* The `flexberry-objectlistview` component:
+    * The `hierarchyAttribute` property.
+    * Default test selector.
+    * Button to clear the filter in a column.
+* The `flexberry-lookup` component:
+    * Support for component's block form, which was previously only available in the mobile version.
+    * The `lookupOnChange` event in the `LookupEvents` service, called when the value in the component changes.
+    * The `autocompleteDirection` property to set direction of autocomplete window.
+* Support of `errors` property change for `flexberry-validationsummary`.
+* The `indeterminate` mode for `flexberry-checkbox` by setting `isNullable`.
+* The `isDesktop` and `isMobile` methods for `device` service.
+* Ability to define default user settings for the type of models to be used in the lists.
+* The `hasShadow` and `hasBorder` properties for `flexberry-toggler` component.
+* The `validationObject` property, `validate` and `saveModel` methods for `edit-form` controller.
 * The localization of the `loading` template.
-* Button to clear the filter in a column in the `flexberry-objectlistview` and `flexberry-simpleolv` components.
-* Default test selector for `flexberry-objectlistview` component.
 
 ### Fixed
+* The `device` service:
+    * `isTv` method.
+* The `edit-form-new` route:
+    * Returning promise in model.
+* The `default-user-setting` blueprint:
+    * Init component path.
+* The `flexberry-objectlistview` component:
+    * Check all rows with respect to `rowConfig.canBeSelected` option.
+    * Add ability to on and off hierarchical mode when component is in hierarchical mode initially.
+* Show save notifications on the edit form.
 * The `updateLookupAction` action in the `flexberry-lookup` component is not called when a value is selected in a modal window.
 * The `index` route is no longer generated in the `flexberry-application` blueprint. Because of this, some of the application regeneration methods did not work.
 * In the row click handler, in the `flexberry-objectlistview` component, async operations outside the ember run loop were used. Now you can just do `click(row)` in your tests.
 
 ### Changed
-* Update dependency on `ember-flexberry-data` to version `2.6.0-beta.0`.
+* Replace `semantic-ui-daterangepicker` dependency with `daterangepicker` 3.0.5.
+* Replace test models localization with separated classes.
 * Now the `flexberry-objectlistview` component is generated with the `showCheckBoxInRow` option set to `true` value on the list form.
 * The `flexberry-acceptance-test` blueprint:
     * An `[AGAT]` tag is added to the name of each test module, which can be used to run the desired test suite. For example `ember test --filter="AGAT"`.
@@ -43,6 +62,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * The `flexberry-lookup` component:
     * Removed mobile template.
     * An exception will be thrown when trying to use the block form of the component with `autocomplete` or `dropdown` options.
+
+## [3.6.0-beta.1] - 2021-04-29
+### Changed
+- Add some improvements for UI of generated application.
+
+## [3.6.0-beta.0] - 2020-10-07
+### Added
+- Changes from [`2.6.0-beta.0`](#260-beta0---2020-10-06) version.
+
+### Changed
+* Update dependency on `ember-flexberry-data` to version `3.6.0-beta.0`.
 
 ## [3.5.0] - 2020-09-17
 ### Added